split tax brackets to another file
